refactor(api): remove model from campaign model
