+++ api/src/prescription/campaign-participation/domain/usecases/send-shared-participation-results-to-pole-emploi.js
@@ -27,8 +27,7 @@ const sendSharedParticipationResultsToPoleEmploi = async ({
   const campaign = await campaignRepository.get(participation.campaignId);
   const organization = await organizationRepository.get(campaign.organizationId);
 
-  //TODO: rewrite when we have only one model for Campaign, for now tests are based on Campaign.js from api context
-  if (campaign.type === 'ASSESSMENT' && organization.isPoleEmploi) {
+  if (campaign.isAssessment && organization.isPoleEmploi) {
     const badges = await badgeRepository.findByCampaignId(participation.campaignId);
     const badgeAcquiredIds = await badgeAcquisitionRepository.getAcquiredBadgeIds({
       badgeIds: badges.map((badge) => badge.id),
