+++ api/src/prescription/campaign/domain/usecases/start-writing-campaign-profiles-collection-results-to-stream.js
@@ -23,7 +23,6 @@ const startWritingCampaignProfilesCollectionResultsToStream = async function ({
   const translate = i18n.__;
   let additionalHeaders = [];
 
-  //TODO: rewrite when we have only one model for Campaign, for now tests are based on Campaign.js from api context
   if (!campaign.isProfilesCollection) {
     throw new CampaignTypeError();
   }
