+++ api/tests/prescription/campaign/integration/infrastructure/repositories/campaign-repository_test.js
@@ -1,10 +1,10 @@
 import _ from 'lodash';
 
+import { Campaign } from '../../../../../../src/prescription/campaign/domain/models/Campaign.js';
 import * as campaignRepository from '../../../../../../src/prescription/campaign/infrastructure/repositories/campaign-repository.js';
 import { CampaignExternalIdTypes } from '../../../../../../src/prescription/shared/domain/constants.js';
 import { CAMPAIGN_FEATURES } from '../../../../../../src/shared/domain/constants.js';
 import { NotFoundError } from '../../../../../../src/shared/domain/errors.js';
-import { Campaign } from '../../../../../../src/shared/domain/models/Campaign.js';
 import { databaseBuilder, domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | Campaign', function () {
