+++ api/tests/tooling/domain-builder/factory/build-campaign.js
@@ -1,7 +1,5 @@
+import { Campaign } from '../../../../src/prescription/campaign/domain/models/Campaign.js';
 import { CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
-import { Campaign } from '../../../../src/shared/domain/models/Campaign.js';
-import { buildOrganization } from './build-organization.js';
-import { buildTargetProfile } from './build-target-profile.js';
 import { buildUser } from './build-user.js';
 
 function buildCampaign({
@@ -18,10 +16,10 @@ function buildCampaign({
   type = CampaignTypes.ASSESSMENT,
   isForAbsoluteNovice = false,
   createdAt = new Date('2020-01-01'),
-  creator = buildUser(),
+  creatorId = buildUser().id,
   ownerId = buildUser().id,
-  organization = buildOrganization(),
-  targetProfile = buildTargetProfile(),
+  organizationId = '18',
+  targetProfileId = '45',
   customResultPageButtonText = null,
   customResultPageButtonUrl = null,
   customResultPageText = null,
@@ -41,10 +39,10 @@ function buildCampaign({
     type,
     isForAbsoluteNovice,
     createdAt,
-    creator,
+    creatorId,
     ownerId,
-    organization,
-    targetProfile,
+    organizationId,
+    targetProfileId,
     customResultPageButtonText,
     customResultPageButtonUrl,
     customResultPageText,
@@ -65,10 +63,10 @@ buildCampaign.ofTypeAssessment = function ({
   type = CampaignTypes.ASSESSMENT,
   isForAbsoluteNovice = false,
   createdAt = new Date('2020-01-01'),
-  creator = buildUser(),
+  creatorId = buildUser().id,
   ownerId = buildUser().id,
-  organization = buildOrganization(),
-  targetProfile = buildTargetProfile(),
+  organizationId = '18',
+  targetProfileId = '45',
   customResultPageButtonText = null,
   customResultPageButtonUrl = null,
   customResultPageText = null,
@@ -87,10 +85,10 @@ buildCampaign.ofTypeAssessment = function ({
     type,
     isForAbsoluteNovice,
     createdAt,
-    creator,
+    creatorId,
     ownerId,
-    organization,
-    targetProfile,
+    organizationId,
+    targetProfileId,
     customResultPageButtonText,
     customResultPageButtonUrl,
     customResultPageText,
@@ -110,9 +108,9 @@ buildCampaign.ofTypeProfilesCollection = function ({
   type = CampaignTypes.PROFILES_COLLECTION,
   isForAbsoluteNovice = false,
   createdAt = new Date('2020-01-01'),
-  creator = buildUser(),
+  creatorId = buildUser().id,
   ownerId = buildUser().id,
-  organization = buildOrganization(),
+  organizationId = '18',
   customResultPageButtonText = null,
   customResultPageButtonUrl = null,
   customResultPageText = null,
@@ -131,9 +129,9 @@ buildCampaign.ofTypeProfilesCollection = function ({
     isForAbsoluteNovice,
     title: null,
     createdAt,
-    creator,
+    creatorId,
     ownerId,
-    organization,
+    organizationId,
     targetProfile: null,
     customResultPageButtonText,
     customResultPageButtonUrl,
